Move computation of PID values to dedicated tasks
